feat: 중복 속성들 분리
